pyproject.toml
[project]
name = "large-netlogo-pairs-dataset"
version = "0.1.0"
description = "Dataset for NetLogo pairs"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "litellm>=1.61.16",
    "python-dotenv>=1.0.1",
]
